Update based on Danny's comment
