--- hudi-client/hudi-spark-client/src/main/java/org/apache/hudi/table/action/commit/SparkInsertOverwriteSimpleBucketIndexPartitioner.java
@@ -1,46 +1,0 @@
-/*
- * Licensed to the Apache Software Foundation (ASF) under one
- * or more contributor license agreements.  See the NOTICE file
- * distributed with this work for additional information
- * regarding copyright ownership.  The ASF licenses this file
- * to you under the Apache License, Version 2.0 (the
- * "License"); you may not use this file except in compliance
- * with the License.  You may obtain a copy of the License at
- *
- *      http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package org.apache.hudi.table.action.commit;
-
-import org.apache.hudi.common.engine.HoodieEngineContext;
-import org.apache.hudi.config.HoodieWriteConfig;
-import org.apache.hudi.index.bucket.BucketIdentifier;
-import org.apache.hudi.table.HoodieTable;
-import org.apache.hudi.table.WorkloadProfile;
-
-/**
- * Packs incoming records to be inserted into buckets (1 bucket = 1 RDD partition).
- */
-public class SparkInsertOverwriteSimpleBucketIndexPartitioner<T> extends SparkBucketIndexPartitioner<T> {
-
-  public SparkInsertOverwriteSimpleBucketIndexPartitioner(
-      WorkloadProfile profile,
-      HoodieEngineContext context,
-      HoodieTable table,
-      HoodieWriteConfig config) {
-    super(profile, context, table, config);
-  }
-
-  @Override
-  public BucketInfo getBucketInfo(int bucketNumber) {
-    String partitionPath = partitionPaths.get(bucketNumber / numBuckets);
-    String bucketId = BucketIdentifier.bucketIdStr(bucketNumber % numBuckets);
-    return new BucketInfo(BucketType.INSERT, BucketIdentifier.newBucketFileIdPrefix(bucketId), partitionPath);
-  }
-}
